+++ lib/ReactViews/Map/Panels/SharePanel/SharePanel.tsx
@@ -12,6 +12,7 @@ import { canShorten } from "./BuildShareLink";
 import Styles from "./share-panel.scss";
 import { SharePanelContent } from "./SharePanelContent";
 import { ShareUrl } from "./ShareUrl";
+import withControlledVisibility from "../../../HOCs/withControlledVisibility";
 
 const MenuPanel =
   require("../../../StandardUserInterface/customizable/MenuPanel").default;
@@ -19,13 +20,13 @@ const StorySharePanel = require("./StorySharePanel").default;
 
 interface PropTypes extends WithTranslation {
   terria: Terria;
-  storyShare: boolean;
+  storyShare?: boolean;
   catalogShare?: boolean;
   catalogShareWithoutText?: boolean;
-  modalWidth: number;
+  modalWidth?: number;
   viewState: ViewState;
-  onUserClick: () => void;
-  btnDisabled: boolean;
+  onUserClick?: () => void;
+  btnDisabled?: boolean;
   t: TFunction;
 }
 
@@ -184,7 +185,7 @@ class SharePanel extends React.Component<PropTypes, SharePanelState> {
   }
 }
 
-export default withTranslation()(SharePanel);
+export default withControlledVisibility(withTranslation()(SharePanel))
 
 export function shouldShorten(terria: Terria) {
   return (