+++ CHANGES.md
@@ -6,6 +6,7 @@
   - getFeaturesFromPickResult now async to handle I3SNode.loadFields()
   - extract common style logic to new Cesium3dTilesStyleMixin.ts
 - Set default value for date and datetime WPS fields only when the field is marked as required.
+- TSify MenuBar
 - [The next improvement]
 
 #### 8.7.5 - 2024-06-26
@@ -15,7 +16,6 @@
 - Show rectangle selector for WPS bounding box parameter
 - Fix `store` and `status` values send in WPS Execute request.
 - Add docs for `modelDimensions`
-- [The next improvement]
 
 #### 8.7.4 - 2024-06-07